refactor: Separate UI/business logic in Todo component
