+++ components/DailyTodos.jsx
@@ -3,12 +3,10 @@ import { CategoryContext } from '@/contexts/CategoryContext';
 import { DateContext } from '@/contexts/DateContext';
 import { LoginContext } from '@/contexts/LoginContext';
 import useTodoStore from '@/contexts/TodoStore';
-import {
-  useTodoAddMutation,
-  useTodoUpdateMutation,
-} from '@/hooks/api/useTodoMutations';
-import { default as useTodosQuery } from '@/hooks/api/useTodoQuery';
-import { isTodoIncludedInTodayView } from '@/utils/dateUtils';
+import useTodosQuery from '@/hooks/api/useTodoQuery';
+import useCreateTodo from '@/hooks/todo/useCreateTodo';
+import useFilteredTodos from '@/hooks/todo/useFilteredTodo';
+import useHandleDrag from '@/hooks/todo/useHandleDrag';
 import {
   DEFAULT_SCROLL_EVENT_THROTTLE,
   handleScroll,
@@ -19,8 +17,7 @@ import {
   TODAYVIEW_TEXTINPUT_SUBMIT_EVENT,
 } from '@/utils/logEvent';
 import { Input } from '@ui-kitten/components';
-import { LexoRank } from 'lexorank';
-import { Fragment, useContext, useEffect, useState } from 'react';
+import { Fragment, useContext } from 'react';
 import { KeyboardAvoidingView, Text, View } from 'react-native';
 import DraggableFlatList, {
   ScaleDecorator,
@@ -29,40 +26,26 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { KeyboardAccessoryView } from 'react-native-keyboard-accessory';
 
 const DailyTodos = () => {
-  const [input, setInput] = useState('');
   const { userId, accessToken } = useContext(LoginContext);
   const { selectedCategory } = useContext(CategoryContext);
-  const { mutate: updateTodo } = useTodoUpdateMutation();
-  const setCurrentTodos = useTodoStore(state => state.setCurrentTodos);
-  const ExistingOrders = useTodoStore(state => state.ExistingOrders);
-  const { mutate: addTodo } = useTodoAddMutation();
   const { selectedDate } = useContext(DateContext);
-  const {
-    isLoading,
-    error,
-    data,
-    isSuccess: isTodosQuerySuccess,
-  } = useTodosQuery(accessToken, userId);
-  const currentTodos = useTodoStore(state => state.currentTodos);
+  const { isLoading, error, data: todosData } = useTodosQuery(userId);
 
-  useEffect(() => {
-    if (isTodosQuerySuccess) {
-      useTodoStore.setState({ todos: data });
-      let filteredTodos = data.filter(
-        todo =>
-          todo.categoryId === selectedCategory &&
-          isTodoIncludedInTodayView(
-            todo.startDate,
-            todo.endDate,
-            selectedDate.format('YYYY-MM-DD'),
-          ),
-      );
-      useTodoStore.setState({ currentTodos: filteredTodos });
-    }
-  }, [isTodosQuerySuccess, data, selectedCategory, selectedDate]);
+  const currentTodos = useFilteredTodos(
+    todosData,
+    selectedCategory,
+    selectedDate,
+  );
+  const setCurrentTodos = useTodoStore(state => state.setCurrentTodos);
 
-  if (isLoading) return <Text>Loading...</Text>;
-  if (error) return <Text>Error: {error.message}</Text>;
+  const handleDragEnd = useHandleDrag(currentTodos, setCurrentTodos);
+
+  const { input, setInput, handleSubmit } = useCreateTodo(
+    userId,
+    accessToken,
+    selectedCategory,
+    selectedDate,
+  );
 
   const renderTodo = ({ item, drag, isActive }) => {
     return (
@@ -71,92 +54,17 @@ const DailyTodos = () => {
       </ScaleDecorator>
     );
   };
-  // const { date } = useContext(DateContext);
 
-  const handleDragEnd = ({ from, to, data: newData }) => {
-    if (!newData || newData.length === 0) return;
-    if (from === to) {
-      return;
-    }
-
-    let updatedOrder;
-    let prevTodoId = null;
-    let nextTodoId = null;
-
-    const ensureUniqueOrder = order => {
-      let uniqueOrder = order;
-      while (true) {
-        if (!ExistingOrders.includes(uniqueOrder.toString())) {
-          break;
-        }
-        if (to === 0) {
-          uniqueOrder = uniqueOrder.genPrev();
-        } else if (to === newData.length - 1) {
-          uniqueOrder = uniqueOrder.genNext();
-        } else {
-          const lexoOrderPrev = uniqueOrder;
-          const lexoOrderNext = LexoRank.parse(newData[to + 1].order);
-          uniqueOrder = lexoOrderPrev.between(lexoOrderNext);
-        }
-      }
-      return uniqueOrder.toString();
-    };
-
-    const orderedNewData = newData.map((item, index) => {
-      if (index !== to) {
-        return item;
-      }
-      let proposedOrder;
-      if (to === 0) {
-        proposedOrder = LexoRank.parse(newData[to + 1].order).genPrev();
-        nextTodoId = newData[to + 1].id;
-      } else if (to === newData.length - 1) {
-        proposedOrder = LexoRank.parse(newData[to - 1].order).genNext();
-        prevTodoId = newData[to - 1].id;
-      } else {
-        const lexoOrderPrev = LexoRank.parse(newData[to - 1].order);
-        const lexoOrderNext = LexoRank.parse(newData[to + 1].order);
-        proposedOrder = lexoOrderPrev.between(lexoOrderNext);
-        prevTodoId = newData[to - 1].id;
-        nextTodoId = newData[to + 1].id;
-      }
-      updatedOrder = ensureUniqueOrder(proposedOrder);
-      return {
-        ...item,
-        order: updatedOrder,
-      };
+  const handleInputSubmit = () => {
+    handleLogEvent(TODAYVIEW_TEXTINPUT_SUBMIT_EVENT, {
+      time: new Date().toISOString(),
+      userId: userId,
     });
-    setCurrentTodos(orderedNewData);
-    const updatedData = {
-      todo_id: newData[to].id,
-      order: {
-        prev_id: prevTodoId,
-        next_id: nextTodoId,
-        updated_order: updatedOrder,
-      },
-    };
-    updateTodo({ accessToken, updatedData });
+    handleSubmit();
   };
 
-  const handleSubmit = async () => {
-    const todos = useTodoStore.getState().todos;
-    const newTodoData = {
-      userId: parseInt(userId, 10),
-      startDate: selectedDate.format('YYYY-MM-DD'),
-      endDate: selectedDate.format('YYYY-MM-DD'),
-      content: input,
-      categoryId: selectedCategory,
-      order:
-        todos.length > 0
-          ? LexoRank.parse(todos[todos.length - 1].order)
-              .genNext()
-              .toString()
-          : LexoRank.middle().toString(),
-    };
-
-    addTodo({ accessToken, todoData: newTodoData });
-    setInput('');
-  };
+  if (isLoading) return <Text>Loading...</Text>;
+  if (error) return <Text>Error: {error.message}</Text>;
 
   return (
     <GestureHandlerRootView style={{ flex: 1 }}>
@@ -182,13 +90,7 @@ const DailyTodos = () => {
               placeholder="새로운 할 일을 입력해주세요"
               value={input}
               onChangeText={setInput}
-              onSubmitEditing={() => {
-                handleLogEvent(TODAYVIEW_TEXTINPUT_SUBMIT_EVENT, {
-                  time: new Date().toISOString(),
-                  userId: userId,
-                });
-                handleSubmit();
-              }}
+              onSubmitEditing={handleInputSubmit}
             />
           </View>
         </KeyboardAccessoryView>
